integrate training mode into flashcards page
